+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Refactored awaitShapeYobta
 - Refactored shapeYobta
 - booleanYobta is now OptionalSyncRule
+- constYobta is now OptionalSyncRule
 
 ## 0.1.6